+++ CHANGELOG.md
@@ -24,6 +24,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\json\isValidJsonDecodeFlags
 - Adds oihana\core\json\isValidJsonEncodeFlags
 
+- Adds oihana\core\maths\gcd
+
 - Adds oihana\core\strings\append
 - Adds oihana\core\strings\block
 - Adds oihana\core\strings\blockPrefix